+++ api/src/error.rs
@@ -37,6 +37,26 @@ pub enum Error {
     #[error("invalid bbox ({0:?}): {1}")]
     InvalidBbox(Vec<f64>, &'static str),
 
+    /// [http::header::InvalidHeaderName]
+    #[error(transparent)]
+    #[cfg(feature = "client")]
+    InvalidHeaderName(#[from] http::header::InvalidHeaderName),
+
+    /// [http::header::InvalidHeaderValue]
+    #[error(transparent)]
+    #[cfg(feature = "client")]
+    InvalidHeaderValue(#[from] http::header::InvalidHeaderValue),
+
+    /// [http::method::InvalidMethod]
+    #[error(transparent)]
+    #[cfg(feature = "client")]
+    InvalidMethod(#[from] http::method::InvalidMethod),
+
+    /// [tokio::task::JoinError]
+    #[error(transparent)]
+    #[cfg(feature = "client")]
+    Join(#[from] tokio::task::JoinError),
+
     /// [std::num::ParseIntError]
     #[error(transparent)]
     ParseIntError(#[from] std::num::ParseIntError),
@@ -44,6 +64,11 @@ pub enum Error {
     /// [std::num::ParseFloatError]
     #[error(transparent)]
     ParseFloatError(#[from] std::num::ParseFloatError),
+
+    /// [reqwest::Error]
+    #[error(transparent)]
+    #[cfg(feature = "client")]
+    Reqwest(#[from] reqwest::Error),
 
     /// A search has both bbox and intersects.
     #[error("search has bbox and intersects")]